[build] Upgrade the MMF SDK from V4.0.0 to V4.1.0 	1. Delete some useless files 	2. Add .gitignore file 	3. add cv181x board configs

+++ build/scripts/sync_git_ver.py
@@ -4,9 +4,11 @@ fd = open("git_version.txt")
 dataList = fd.readlines()
 rootdir = os.getcwd()
 for i in range(0, len(dataList), 3):
-    os.chdir("{location}".format(location=dataList[i].split(' ')[1].strip()))
-    os.system("git reset --hard {commit}".format(
-        commit=dataList[i + 1].split(' ')[0])
-    )
-    os.chdir(rootdir)
+    gitDir = "{location}".format(location=dataList[i].split(' ')[1].strip())
+    if os.path.exists(gitDir):
+        os.chdir(gitDir)
+        os.system("git reset --hard {commit}".format(
+            commit=dataList[i + 1].split(' ')[0])
+        )
+        os.chdir(rootdir)
 fd.close()
